+++ src/Exception/InvalidCallException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Bupy7\Form\Exception;
+
+use BadMethodCallException;
+
+/**
+ * @author Belosludcev Vasily <https://github.com/bupy7>
+ * @since 2.0.0
+ */
+class InvalidCallException extends BadMethodCallException
+{
+}