+++ src/Exception/UnknownPropertyException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Bupy7\Form\Exception;
+
+use Exception;
+
+/**
+ * @author Belosludcev Vasily <https://github.com/bupy7>
+ * @since 2.0.0
+ */
+class UnknownPropertyException extends Exception
+{
+}