+++ tests/Asset/EmptyForm.php
@@ -5,11 +5,9 @@ namespace Bupy7\Form\Tests\Asset;
 use Bupy7\Form\FormAbstract;
 
 /**
- * Empty form.
  * @author Belosludcev Vasilij <https://github.com/bupy7>
  * @since 1.0.0
  */
 class EmptyForm extends FormAbstract
 {
-
 }