+++ tests/Asset/SignInFilter.php
@@ -5,15 +5,11 @@ namespace Bupy7\Form\Tests\Asset;
 use Zend\InputFilter\InputFilter;
 
 /**
- * Input filter of signin.
  * @author Belosludcev Vasilij <https://github.com/bupy7>
  * @since 1.0.0
  */
 class SignInFilter extends InputFilter
 {
-    /**
-     * {@inheritDoc}
-     */
     public function __construct()
     {
         $this->add([