+++ tests/ErrorStorage/ErrorStorageTest.php
@@ -3,7 +3,7 @@
 namespace Bupy7\Form\Tests\ErrorStorage;
 
 use PHPUnit_Framework_TestCase;
-use Bupy7\Form\Tests\Asset\SignInForm;
+use Bupy7\Form\Tests\Asset\SignInPropertyForm;
 use Bupy7\Form\ErrorStorage\ErrorStorage;
 
 /**
@@ -17,7 +17,7 @@ class FormAstractTest extends PHPUnit_Framework_TestCase
      */
     public function testErrorHints()
     {
-        $signInForm = new SignInForm;
+        $signInForm = new SignInPropertyForm;
         $signInForm->setValues([
             'email' => 'test@gmail.co2m',
             'password' => '12q34e56t78',
@@ -31,4 +31,3 @@ class FormAstractTest extends PHPUnit_Framework_TestCase
         $this->assertEquals(null, $errorStorage->getError('password'));
     }
 }
-