+++ tests/View/Helper/FormBuilderHelperFactoryTest.php
@@ -7,7 +7,7 @@ use Zend\View\HelperPluginManager;
 use Zend\ServiceManager\ServiceManager;
 use AdamWathan\Form\FormBuilder;
 use Bupy7\Form\View\Helper\FormBuilderHelper;
-use Bupy7\Form\Tests\Asset\SignInForm;
+use Bupy7\Form\Tests\Asset\SignInPropertyForm;
 
 /**
  * @author Vasilij Belosludcev <https://github.com/bupy7>
@@ -23,7 +23,7 @@ class FormBuilderHelperFactoryTest extends PHPUnit_Framework_TestCase
         $moduleConfig = require __DIR__ . '/../../../config/module.config.php';
         $serviceManager = new ServiceManager($moduleConfig['service_manager']);
         $helperPluginManager = new HelperPluginManager($serviceManager, $moduleConfig['view_helpers']);
-        $signInForm = new SignInForm;
+        $signInForm = new SignInPropertyForm;
 
         $formBuilderHelper1 = $helperPluginManager->get('formBuilder');
         $formBuilderHelper1()->setToken('test token 1');
@@ -34,7 +34,5 @@ class FormBuilderHelperFactoryTest extends PHPUnit_Framework_TestCase
         $this->assertInstanceOf(FormBuilder::class, $formBuilderHelper2($signInForm));
     
         $this->assertNotEquals($formBuilderHelper1(), $formBuilderHelper2());
-
     }
 }
-