+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 zf-form
 =======
 
+2.0.0
+-----
+
+- Enh: Rewrite tests.
+- Add: Scenarios.
+- Add: Property or method fields.
+
 1.0.0 [2016-08-28]
 ------------------